% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateArLog.R
\name{create_arlog}
\alias{create_arlog}
\title{Create and Combine Data Frames with Date Processing and Join}
\usage{
create_arlog(CSV_OUTPUT_DIR, dfs_for_mod, pdf_cpo_seq_processed)
}
\arguments{
\item{CSV_OUTPUT_DIR}{A character string representing the directory path where
the output will be saved (currently not used in the function, but can be
included for future extensions).}

\item{dfs_for_mod}{A list of data frames that need to be processed. The date
columns (\code{date_of_birth} and \code{date_of_collection}) in these data frames
will be standardized to a common date format (\code{ymd}).}

\item{pdf_cpo_seq_processed}{A data frame that contains the \code{COLLECTION_DATE}
column to be joined with the \code{date_of_collection} column in the combined data
frame.}
}
\value{
A data frame that is the result of the left join between the combined
\code{dfs_for_mod} data frames and \code{pdf_cpo_seq_processed}, with the common columns
\code{arln_accession_id} and \code{date_of_collection} properly aligned.
}
\description{
This function processes a list of data frames, performs date conversions, and
combines them into a single data frame. It then joins the combined data frame
with another processed data frame (\code{pdf_cpo_seq_processed}) based on common
columns (\code{arln_accession_id} and \code{date_of_collection}).
}
\details{
This function first processes the list of data frames provided in
\code{dfs_for_mod}, converting the \code{date_of_collection} column to a consistent
\code{Date} format using \code{lubridate::ymd()}. The processed data frames are then
combined into a single data frame. After that, the function performs a left
join with the \code{pdf_cpo_seq_processed} data frame, based on the \code{arln_accession_id}
and \code{date_of_collection} columns.
}
\examples{
# Example usage
processed_dfs <- list(df1, df2, df3)  # Your list of data frames
pdf_cpo_seq <- data.frame(ACCESSION_ID = c(1, 2, 3),
                          COLLECTION_DATE = as.Date(c("2021-01-01", "2021-01-02", "2021-01-03")))
result <- create_arlog(CSV_OUTPUT_DIR = "output_folder", dfs_for_mod = processed_dfs, pdf_cpo_seq_processed = pdf_cpo_seq)

}
\seealso{
\code{\link{dplyr::bind_rows}}, \code{\link{lubridate::ymd}}, \code{\link{dplyr::left_join}}
}
