% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsCreateArlog.R
\name{process_dates}
\alias{process_dates}
\title{Process Date Columns in a List of Data Frames}
\usage{
process_dates(dataframes)
}
\arguments{
\item{dataframes}{A list of data frames that need to be processed. Each data frame
should contain \code{date_of_birth} and \code{date_of_collection} columns that will be
converted to \code{Date} objects.}
}
\value{
A list of data frames with the \code{date_of_birth} and \code{date_of_collection} columns
converted to Date format. The input list is modified in place.
}
\description{
This function takes a list of data frames and standardizes the format of the
\code{date_of_birth} and \code{date_of_collection} columns to a consistent Date format.
It ensures that all date columns in the data frames are converted to the
specified format (\verb{\%m/\%d/\%Y}), which is typically used for mm/dd/yyyy formatted dates.
}
\examples{
# Example usage with a list of data frames
dataframes <- list(ta, wa, ecp, esp, ewp)
processed_dataframes <- process_dates(dataframes)

# Access processed data frames
processed_dataframes[[1]]  # ta
processed_dataframes[[2]]  # wa

}
\seealso{
\code{\link{as.Date}}, \code{\link{bind_rows}}
}
