% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessJoinedFiles.R
\name{process_tenn_arln}
\alias{process_tenn_arln}
\title{Process the 'tenn_arln' Data Frame and Save as CSV}
\usage{
process_tenn_arln(tenn_arln_df)
}
\arguments{
\item{tenn_arln_df}{A data frame with columns that may contain spaces, hash characters,
and uncapitalized letters in the column names.}
}
\value{
A data frame with modified column names: spaces replaced by underscores, hash characters replaced by 'num', and capitalized column names.
}
\description{
This function processes the \code{tenn_arln_df} data frame by applying a series of
transformations:
\enumerate{
\item Replaces spaces in column names with underscores.
\item Replaces any \verb{#} characters in column names with the string "Num".
\item Capitalizes all column names.
}
}
\examples{
# Example usage:
processed_df <- process_tenn_arln(tenn_arln_df)

}
