% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessJoinedFiles.R
\name{process_word_alert}
\alias{process_word_alert}
\title{Process the 'word_alert' Data Frame}
\usage{
process_word_alert(word_alert_df)
}
\arguments{
\item{word_alert_df}{A data frame with columns that may contain spaces, hyphens,
and need to have certain names replaced (e.g., DOB).}
}
\value{
A processed data frame with:
\itemize{
\item Column names with spaces replaced by underscores and capitalized.
\item "DOB" replaced with "DATE_OF_BIRTH".
\item Hyphens in the data replaced with NA.
\item Date columns converted to the format specified.
}
}
\description{
This function processes the \code{word_alert_df} data frame by applying a series of
transformations:
\enumerate{
\item Replaces spaces in column names with underscores.
\item Capitalizes all column names.
\item Replaces "DOB" with "DATE_OF_BIRTH" in column names.
\item Replaces hyphens (more than 2 consecutive hyphens) in character columns
with \code{NA}.
\item Converts date columns ("DATE_OF_BIRTH" and "DATE_OF_COLLECTION") from
character to Date type using the specified date format (\verb{\%m-\%d-\%Y}).
}
}
\examples{
# Example usage:
processed_df <- process_word_alert(word_alert_df)

}
