% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArlogTransformers.R
\name{transform_excel_sent}
\alias{transform_excel_sent}
\title{Transform Excel Sent Data}
\usage{
transform_excel_sent(df)
}
\arguments{
\item{df}{A dataframe containing columns related to Excel Sent data.}
}
\value{
A transformed dataframe with the same structure as the desired output.
}
\description{
This function transforms the column names of a dataframe to match the structure
of the SQL query format for Excel Sent data, ensuring all columns are accounted for.
}
\examples{
df <- data.frame(
  DISEASE = c("Disease1", "Disease2"),
  ADMINSTATUS = c("Active", "Inactive"),
  PATIENTID = c("P123", "P456"),
  INCIDENTID = c("I123", "I456"),
  LASTNAME = c("Doe", "Smith"),
  FIRSTNAME = c("John", "Jane"),
  DATE_OF_BIRTH = as.Date(c("1990-01-01", "1985-05-15")),
  LAB = c("Lab1", "Lab2"),
  LABOTHER = c("OtherLab1", "OtherLab2"),
  ACCESSION_NUM = c("A123", "A456"),
  CULTUREDT = as.Date(c("2024-01-01", "2024-02-01")),
  ORGANISM = c("Organism1", "Organism2"),
  SPECIMENTYPE = c("Type1", "Type2"),
  CULTURE...14 = c("Culture1", "Culture2"),
  PHENOTYPIC = c("Positive", "Negative"),
  CIMRESULT = c("Resistant", "Sensitive"),
  MCIMRESULT = c("Positive", "Negative"),
  MHTRESULT = c("Positive", "Negative"),
  MBLRESULT = c("Positive", "Negative"),
  CARBANPRESULT = c("Positive", "Negative"),
  GENOTYPIC = c("Gene1", "Gene2"),
  PCRRESULT = c("Positive", "Negative"),
  CARBARRESULT = c("Positive", "Negative"),
  CULTURE...24 = c("Culture3", "Culture4"),
  CULTUREFAC = c("Facility1", "Facility2"),
  HOSPFAC = c("Hospital1", "Hospital2"),
  OUTBREAKRELATE = c("Yes", "No"),
  OTHEROUTBREAK = c("No", "Yes"),
  COLLECTIONDT = as.Date(c("2024-01-01", "2024-02-01")),
  ORG = c("Org1", "Org2"),
  SPECIMENSOURCE = c("Blood", "Saliva"),
  CULTURE = c("Culture5", "Culture6")
)
transformed_df <- transform_excel_sent(df)
}
