% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArlogTransformers.R
\name{transform_web_portal}
\alias{transform_web_portal}
\title{Transform Web Portal Data}
\usage{
transform_web_portal(df)
}
\arguments{
\item{df}{A dataframe with the appropriate columns from the web portal data.}
}
\value{
A transformed dataframe with the same structure as the SQL query.
}
\description{
This function transforms the column names of an R dataframe to match the structure
specified in the given SQL query for the \code{public.web_portal} table.
}
\examples{
df <- data.frame(PATIENT_NAME = c("John Doe", "Jane Smith"),
                 PATIENT_DOB = as.Date(c("1990-01-01", "1985-05-15")),
                 LIMS_ACCESSION_ID = c("L123", "L456"),
                 DATE_COLLECTED = as.Date(c("2024-01-01", "2024-02-02")),
                 DATE_RECEIVED = as.Date(c("2024-01-02", "2024-02-03")),
                 FACILITY_NAME = c("Test Lab 1", "Test Lab 2"))
transformed_df <- transform_web_portal(df)
}
