% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArlogTransformers.R
\name{transform_word_alert}
\alias{transform_word_alert}
\title{Transform Word Alert Data}
\usage{
transform_word_alert(df)
}
\arguments{
\item{df}{A dataframe containing columns related to Word Alert data.}
}
\value{
A transformed dataframe with the same structure as the desired output.
}
\description{
This function transforms the column names of a dataframe to match the structure
of the SQL query format for Word Alert data.
}
\examples{
df <- data.frame(PATIENT_NAME = c("John Doe", "Jane Smith"),
                 LAB_ID = c("L123", "L456"),
                 DATE_OF_BIRTH = as.Date(c("1990-01-01", "1985-05-15")),
                 RECEIVED_FROM = c("Facility A", "Facility B"),
                 SOURCE = c("Blood", "Saliva"),
                 DATE_OF_COLLECTION = as.Date(c("2024-01-01", "2024-02-02")),
                 ORGANISM_ID = c("Org1", "Org2"),
                 CARBA_R_RESULT = c("Positive", "Negative"))
transformed_df <- transform_word_alert(df)
}
