% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputControl.R
\name{write_dfs_to_csv}
\alias{write_dfs_to_csv}
\title{Write Multiple DataFrames to CSV Files with Custom Filenames}
\usage{
write_dfs_to_csv(dfs, output_dir, filenames)
}
\arguments{
\item{dfs}{A list of data frames to be written to CSV files.}

\item{output_dir}{A character string indicating the directory where the CSV files should be saved.}

\item{filenames}{A character vector containing the names for the output CSV files. The length must match the number of data frames in \code{dfs}.}
}
\value{
NULL This function writes the CSVs to the provided directory but does not return any value.
}
\description{
This function takes a list of data frames and a corresponding list of filenames, then writes each data frame to a CSV file with the provided filenames.
}
