% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArlogTransformers.R
\name{transform_excel_cpo}
\alias{transform_excel_cpo}
\title{Transform Excel CPO Data}
\usage{
transform_excel_cpo(df)
}
\arguments{
\item{df}{A dataframe containing columns related to Excel CPO data.}
}
\value{
A transformed dataframe with the same structure as the desired output.
}
\description{
This function transforms the column names of a dataframe to match the structure
of the SQL query format for Excel CPO data, ensuring all columns are accounted for
including those with hyphens and special characters.
}
